+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [STREAM-2184](https://inindca.atlassian.net/browse/STREAM-2184) - Update stanza to v12.22.1
 * [STREAM-2176](https://inindca.atlassian.net/browse/STREAM-2176) - Allow userId to be passed in as a client option so it doesn't need to be fetched again
 * [STREAM-2207](https://inindca.atlassian.net/browse/STREAM-2207) - Debounce `claimAlertingLeader` to prevent hammering the endpoint
+* [STREAM-1095](https://inindca.atlassian.net/browse/STREAM-1095) - Set `connected` just before emitting the `connected` event to avoid a gap where an extension might see `connected` as `true` but not have the current stanza instance internally yet.
 
 # [v20.0.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v20.0.0...v20.0.1)
 ### Changed